+++ src/js/pages/Home.js
@@ -1,0 +1,13 @@
+import { Col, Row } from "react-bootstrap";
+import HomePresentational from "../components/presentationals/HomePresentational"
+
+const Home = () => (
+    <Row>
+        <Col>
+            <HomePresentational />
+        </Col>
+    </Row>
+);
+
+
+export default Home;